requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
docopt==0.6.2
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
pipreqs==0.4.13
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.6
Wikipedia-API==0.6.0
yarg==0.1.9
